test: extract turn generation white-box suite
